+++ tests/dummy/app/JsonApi/Posts/Validators.php
@@ -18,6 +18,8 @@
 namespace DummyApp\JsonApi\Posts;
 
 use CloudCreativity\LaravelJsonApi\Rules\DateTimeIso8601;
+use CloudCreativity\LaravelJsonApi\Rules\HasMany;
+use CloudCreativity\LaravelJsonApi\Rules\HasOne;
 use CloudCreativity\LaravelJsonApi\Validation\AbstractValidators;
 use DummyApp\Post;
 
@@ -114,8 +116,8 @@ class Validators extends AbstractValidators
                 'nullable',
                 new DateTimeIso8601()
             ],
-            'author.type' => 'in:users',
-            'tags.*.type' => 'in:tags',
+            'author' => new HasOne('users'),
+            'tags' => new HasMany('tags'),
         ];
     }